#45: Added the Money class and changed price type from BigDecimal to String

+++ src/main/java/com/blinenterprise/SyropKlonowy/repository/ProductRepository.java
@@ -5,7 +5,6 @@ import com.blinenterprise.SyropKlonowy.domain.Product;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,13 +17,13 @@ public interface ProductRepository extends CrudRepository<Product, Long> {
 
     List<Product> findAllByCategory(Category category);
 
-    List<Product> findAllByPriceLessThanEqual(BigDecimal price);
+    List<Product> findAllByPriceLessThanEqual(float price);
 
-    List<Product> findAllByPriceGreaterThanEqual(BigDecimal price);
+    List<Product> findAllByPriceGreaterThanEqual(float price);
 
-    List<Product> findAllByPriceLessThanEqualAndCategory(BigDecimal price, Category category);
+    List<Product> findAllByPriceLessThanEqualAndCategory(float price, Category category);
 
-    List<Product> findAllByPriceGreaterThanEqualAndCategory(BigDecimal price, Category category);
+    List<Product> findAllByPriceGreaterThanEqualAndCategory(float price, Category category);
 
     Optional<Product> findByCode(String code);
 }
